FIX : Replacing the try / catch by if / alse as done previously and replacing the .then in isLiked function by a proper await.

+++ src/modules/like/services/like.service.ts
@@ -1,7 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { LikeRepository } from '../repositories/like.repository';
 import { AlreadyLikedException } from '../exceptions/already-liked.exception';
-import { Prisma } from '@prisma-generated/client';
 import { LikeDoesNotExistException } from '../exceptions/like-does-not-exist.exception';
 
 @Injectable()
@@ -9,40 +8,23 @@ export class LikeService {
 	constructor(private readonly likeRepository: LikeRepository) {}
 
 	async addLike(postId: string, userId: string) {
-		try {
-			return await this.likeRepository.addLike(postId, userId);
-		} catch (err) {
-			if (
-				err instanceof Prisma.PrismaClientKnownRequestError &&
-				err.code === 'P2002'
-			) {
-				throw new AlreadyLikedException();
-			}
-			throw err;
-		}
+		const isLiked = await this.isLiked(postId, userId);
+		if (isLiked) throw new AlreadyLikedException();
+		await this.likeRepository.addLike(postId, userId);
 	}
 
 	async deleteLike(postId: string, userId: string) {
-		try {
-			return await this.likeRepository.deleteLike(postId, userId);
-		} catch (err) {
-			if (
-				err instanceof Prisma.PrismaClientKnownRequestError &&
-				err.code === 'P2025'
-			) {
-				throw new LikeDoesNotExistException();
-			}
-			throw err;
-		}
+		const isLiked = await this.isLiked(postId, userId);
+		if (!isLiked) throw new LikeDoesNotExistException();
+		await this.likeRepository.deleteLike(postId, userId);
 	}
 
 	countLike(postId: string) {
 		return this.likeRepository.countLike(postId);
 	}
 
-	isLiked(postId: string, userId: string) {
-		return this.likeRepository
-			.isLiked(postId, userId)
-			.then((like) => !!like);
+	async isLiked(postId: string, userId: string) {
+		const like = await this.likeRepository.isLiked(postId, userId);
+		return !!like;
 	}
 }
